--- src/components/Progress/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Progress';
