--- src/domains/Townhall/ActionHistory/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ActionHistory';
